Query visjs graph state to determine actions to take
Allow the user to hand-edit the graph, after search criteria are entered, to further reduce what a campaign attribution is
to be applied to. This attempts to work the existing logic into this new functionality. Some experimentation will be done
to determine if a change to the workflow for this (and similar) bulk actions can be done differently.